Refactor: standardize naming, schema, and seed strategy
- Renamed key properties for clarity (e.g., codeAirlines → codeAirlinesIcao, codeAirport → codeAirportIcao, etc.)
- Updated entities, DTOs, and services to use new property names
- Changed key types for Permissions, RolPermission, and ProviderResponse to short
- Made Organizations class concrete and mapped inheritance in DbContext
- Updated FlightHistory composite key and related logic
- Switched seed data to use raw SQL for idempotency in migrations
- Adjusted migrations and model snapshots to match new schema
- Suppressed EF Core pending model change warnings
- Performed minor code cleanups and namespace fixes

+++ Core/AeroVeloz.Domain/Entities/Notification/ProviderResponse.cs
@@ -12,7 +12,7 @@ namespace AeroVeloz.Domain.Entities.Notification
 
     [Table("ProviderResponse", Schema = "Notifications")]
 
-    public partial class ProviderResponse: BEntity<byte>
+    public partial class ProviderResponse: BEntity<short>
     {
         public string? name { get; init; }
     }
