+++ Core/AeroVeloz.Domain/Models/Permission/PermissionModel.cs
@@ -1,7 +1,7 @@
 ﻿namespace AeroVeloz.Domain.Models.Permission
 {
     public sealed record PermissionModel(
-         byte idPermission, 
+         short idPermission, 
          string? codePermission
         );
